Moved date utils to dateTimeUtils.ts
