+++ packages/react/src/auto/mui/inputs/MUIAutoDateTimePicker.tsx
@@ -2,8 +2,8 @@ import { Box } from "@mui/material";
 import { DatePicker, TimePicker } from "@mui/x-date-pickers";
 import React from "react";
 import { useController } from "react-hook-form";
+import { zonedTimeToUtc } from "../../../dateTimeUtils.js";
 import type { GadgetDateTimeConfig } from "../../../internal/gql/graphql.js";
-import { zonedTimeToUtc } from "../../../utils.js";
 import { useFieldMetadata } from "../../hooks/useFieldMetadata.js";
 
 export const MUIAutoDateTimePicker = (props: { field: string; value?: Date; onChange?: (value: Date) => void; error?: string }) => {
