+++ packages/react/src/auto/polaris/inputs/PolarisAutoDateTimePicker.tsx
@@ -3,8 +3,8 @@ import { DatePicker, Icon, InlineStack, Popover, TextField } from "@shopify/pola
 import { CalendarIcon } from "@shopify/polaris-icons";
 import React, { useCallback, useMemo, useState } from "react";
 import { useController } from "react-hook-form";
+import { formatShortDateString, isValidDate, utcToZonedTime, zonedTimeToUtc } from "../../../dateTimeUtils.js";
 import type { GadgetDateTimeConfig } from "../../../internal/gql/graphql.js";
-import { formatShortDateString, isValidDate, utcToZonedTime, zonedTimeToUtc } from "../../../utils.js";
 import { useFieldMetadata } from "../../hooks/useFieldMetadata.js";
 import type { DateTimeState } from "./PolarisAutoTimePicker.js";
 import PolarisAutoTimePicker from "./PolarisAutoTimePicker.js";
